# Improve performance of FileSystemFontProvider.scanFonts()