--- a/fontbox/src/main/java/org/apache/fontbox/ttf/TTFTable.java
+++ b/fontbox/src/main/java/org/apache/fontbox/ttf/TTFTable.java
@@ -126,4 +126,16 @@ public class TTFTable
     void read(TrueTypeFont ttf, TTFDataStream data) throws IOException
     {
     }
+    
+    /**
+     * This will read required headers from the stream into outHeaders.
+     * 
+     * @param ttf The font that is being read.
+     * @param data The stream to read the data from.
+     * @param outHeaders The class to write the data to.
+     * @throws IOException If there is an error reading the data.
+     */
+    void readHeaders(TrueTypeFont ttf, TTFDataStream data, FontHeaders outHeaders) throws IOException
+    {
+    }
 }