… fontbox/src/main/java/org/apache/fontbox/ttf/TrueTypeCollection.java
… fontbox/src/main/java/org/apache/fontbox/ttf/TrueTypeFont.java
… pdfbox/src/main/java/org/apache/pdfbox/pdmodel/font/FileSystemFontProvider.java