@PascalSchumacher 2024-05-21
Hi, any update on this? Looks like a worthwhile enhancement (as far as I can tell).