@bogdiuk 2024-06-18
Force-pushed the new implementation, based on 3.0 branch (rev 5c4a09a5c0574c1701151202ede2be1857610a36). Also I edited the 1st post to reflect changes

Performance numbers: first number is a re-run (font files are cached by the OS), second is a cold run

Windows (296 fonts):
- 2.9s..4.0s - `3.0` branch
- 0.8s..1.9s - `performance-scanfonts` branch
- 0.3s..0.6s - `performance-scanfonts` +"pdfbox.fontcache.skipchecksums"

macOS (792 fonts):
- 2.2..3.6s - `3.0` branch
- 0.4..0.5s - `performance-scanfonts` branch
- 0.16s..0.2s `performance-scanfonts` +"pdfbox.fontcache.skipchecksums"