@bogdiuk 2024-06-18
Reasoning behind using a separate class `LoadOnlyHeaders` instead of extracting information directly from `TrueTypeFont`:
* easier to debug - single place to put breakpoints
* easier to refactor:
  * for example, simple "Find All Usages" can reveal that `setSomething()` is not used
  * the alternative is having `boolean isLoadOnlyHeaders` in `TTFParser`, `CFFParser` and `TrueTypeFont`: it would be harder to collect usages of all 3
* in the future, parsing some tables can be skipped completely, only setting respective fields of `LoadOnlyHeaders`