@THausherr 2024-06-25
Re "fixed a memory leak", this is tricky: your change makes sense for the `TrueTypeCollection(File file)` constructor only. Not for the one with InputStream. That one (which isn't used by PDFBox itself) also reads the font into memory twice as it is now.