@bogdiuk 2024-06-25
Thank you, I will fix the issues today.

Re `readExact()`/`readUpTo()`: I will remove them for simplicity, they remained from my initial PR.

I initially saw that `RandomAccessRead.read()` was called in a loop and assumed that it behaves like `InputStream.read()` (i.e. can return `<N` bytes and still have more coming), so it needs to be wrapped into `readUpTo()`.
But now I see that all implementations of `RandomAccessRead.read()` have loops of their own, so they already behave like `InputStream.readNBytes()` (can only read less than N bytes in case of EOF)? But this is not specified in Javadoc.

For the future refactoring, here are the places where `read()` is called in a loop (either shrink them to 1 call or add `readUpTo()` that guarantees `==N bytes || EOF`):
* `RandomAccessReadDataStream.cctor()`
* `COSParser.getStartxrefOffset()` (but also throws `if (readBytes < 1)`)
* `PDFXrefStreamParser.readNextValue()`
* `PDCIDFontType2.getParser()`
* `PDTrueTypeFont.getParser()`
* `PDFontFactory.getFontHeader()`