@THausherr 2024-06-27
I applied the patch for the first time, I do get differences in the `pdfbox.cache` file:
ARBONNIE becomes ARBON
ComicSansMS-BoldItalic becomes Com
GeorgiaPro-CondSemiboldItalic becomes Georgi
NotoNaskhArabicUI-Bold becomes NotoNaskhArabicUI-
KodchiangUPCBoldItalic becomes Kodchi

Here's the comicz file: 
[comicz.zip](https://github.com/user-attachments/files/16012154/comicz.zip)

However it might also be my fault due to my own changes, so I'll investigate.