@bogdiuk 2024-06-27
Re exception: 2 calls should've been `setError(String)` to avoid slow `fillInStackTrace()` in the exception constructor. Third call removed because existing exceptions should be forwarded indeed.

Re "font.getNaming()" - don't remember, looks weird, removed. 

Found a potential bug in `TrueTypeCollection.processAllFontHeaders()` and `processAllFonts()`: if some font in TTC file is corrupted, an exception is thrown and all remaining fonts in TTC are skipped. Didn't fix because it's debatable (maybe whole TTC is corrupted, there's no need to waste time on parsing it).

Also, simplified `addTrueTypeCollection()` because all 3 methods it calls do not effectively throw.