@THausherr 2024-06-29
Thanks; ` IOUtils.closeQuietly(parser.parse(new RandomAccessReadBufferedFile(ttfFile)));` caught my attention because `parser.parse()` returns null. So the closeQuietly() isn't needed. And returning sometimes a font and sometimes null kindof contradicts the "do one thing" ideal. How about making a different method for the headers instead, which returns a FontHeader object, and refactor the common code of the two methods into a third one? Same for `getFontAtIndex`, IMHO this should be split in `getFontAtIndex` and `getFontHeaderAtIndex` and the common part refactored into a method accessed by both, this would be more clear.