@bogdiuk 2024-06-30
While I'm at it, I think it's better to replace "calling getter just for side-effects"
```java
font.getNaming(); // calls NamingTable.readTable();
font.getHeader(); // calls HeaderTable.readTable();
((OpenTypeFont) font).getCFF(); // calls CFFTable.readTable();
```
with explicit calls
```java
font.readTableHeaders(*.TAG, outHeaders); // -> TTFTable.readHeaders(..., FontHeaders)
```
Then `TTFParser.loadOnlyHeaders` field can be removed.

The changeset will become a bit larger (like, duplicated `TrueTypeFont.readTable()`), but codepaths will be more clear