@bogdiuk 2024-06-30
Or did I over-do it with `CFFParser.parseFirstSubFontROS()`?