@bogdiuk 2024-06-30
When `onlyHeaders==true`, it discards useless entries, making `nameRecords` smaller and thus `readString()` is called fewer times in the following loop, reducing both computing and file reading.

So it's slightly more computing in the 1st loop (`isUsefulForOnlyHeaders()` is just 5 comparisons), but much less computing in the second (seek > allocate+read > detect charset > new String) and `fillLookupTable()` that traverse `nameRecords`

Of course, to please branch predictor, we can extract `if` from the loop, but it looks much more verbose and it should be easy to predict for CPU anyway:
```
    if (onlyHeaders)
    {
        for (int i=0; i< numberOfNameRecords; i++)
        {
            NameRecord nr = new NameRecord();
            nr.initData(ttf, data);
            if (isUsefulForOnlyHeaders(nr))
            {
                nameRecords.add(nr);
            }
        }
    } else {
        for (int i=0; i< numberOfNameRecords; i++)
        {
            NameRecord nr = new NameRecord();
            nr.initData(ttf, data);
            nameRecords.add(nr);
        }
    }
```

Which 3 lines do you mean?