@bogdiuk 2024-06-30
As for `lookupTable`, I am going to remove in the the future PR (the change was in my initial Request).
By sorting `nameRecords`, we can use a simple binary search in `getName()`, which is fast enough, and doesn't need additional memory, like current 4-level HashMap cache. Especially considering that each name is searched only once