@THausherr 2024-06-30
I meant these:
```
        lookupTable = new HashMap<>(nameRecords.size());
        fillLookupTable();
        readInterestingStrings();
```